[KYUUBI #3950] Fix the batch metadata in-consistent issue on open batch session failure
### _Why are the changes needed?_

I found that, there is some confused batch metadata.

The metadata state in mysql is PENDING, but the batch session state is `ERROR`.

I think the RC is that:
- the metadata insert is in KyuubiBatchSessionImpl
- If open batch session failed, the BatchJobSubmission opHandle has not been added into batch session operation set.
- then the close session will not update the batch metadata

In this pr, I re-order the code when opening batch session.
### _How was this patch tested?_
- [ ] Add some test cases that check the changes thoroughly including negative and positive cases if possible

- [ ] Add screenshots for manual tests if appropriate

- [x] [Run test](https://kyuubi.apache.org/docs/latest/develop_tools/testing.html#running-tests) locally before make a pull request

Closes #3950 from turboFei/update_batch_state.

Closes #3950

49a8e073c [fwang12] try to kill
fd495bcd7 [fwang12] fix ut
6d3bd69f3 [fwang12] add ut
bb2bf89a7 [fwang12] add more log
0c9f32b68 [fwang12] check recovery metadata if failed before run batch op
23f8fef29 [fwang12] comments
e45b1924c [fwang12] re-order
0f0ad8422 [fwang12] set launchEngineOp to lazy to prevent the OP CONN metrics failure
0eed14dff [fwang12] add batch op
c39da31a2 [fwang12] Revert "insert metadata when run operation"
61b40c468 [fwang12] insert metadata when run operation

Authored-by: fwang12 <fwang12@ebay.com>

+++ kyuubi-server/src/main/scala/org/apache/kyuubi/operation/BatchJobSubmission.scala
@@ -285,6 +285,17 @@ class BatchJobSubmission(
       }
 
       MetricsSystem.tracing(_.decCount(MetricRegistry.name(OPERATION_OPEN, opType)))
+
+      if (!builder.processLaunched) {
+        builder.close()
+        if (recoveryMetadata.isDefined) {
+          killMessage = killBatchApplication()
+        }
+        setState(OperationState.CANCELED)
+        updateBatchMetadata()
+        return
+      }
+
       // fast fail
       if (isTerminalState(state)) {
         killMessage = (false, s"batch $batchId is already terminal so can not kill it.")
